+++ tests/command/mod.rs
@@ -147,6 +147,7 @@ mod diff_test;
 mod fetch_test;
 mod index_pack_test;
 mod init_from_git_test;
+mod init_json_test;
 mod init_separate_libra_dir_test;
 mod init_test;
 mod lfs_test;